+++ inst/include/rxode2parseStruct.h
@@ -623,6 +623,10 @@ typedef struct rx_solve_s {
   int ssInfDurAbort;
   int *splitBolus;
   int splitBolusN;
+  int *splitInfusionBolus;
+  int splitInfusionBolusN;
+  int *splitBolusInfusion;
+  int splitBolusInfusionN;
   rx_fn_pointers fns;
 } rx_solve;